premium_features.py

import discord
from discord.ext import commands
from utils.embeds import create_embed
from datetime import datetime, timedelta

class PremiumFeatures(commands.Cog):
    def __init__(self, bot):
        self.bot = bot
        self.premium_servers = set()  # Gerçek projede veritabanında olacak
        
    def is_premium_server(self, guild_id):
        """Check if server has premium subscription."""
        return guild_id in self.premium_servers
    
    @commands.command()
    @commands.has_permissions(administrator=True)
    async def premium(self, ctx):
        """Show premium features and pricing."""
        embed = create_embed(
            title="💎 IronWard Premium Features",
            description="Botunuzu bir sonraki seviyeye taşıyın!",
            color=discord.Color.gold()
        )
        
        # Free vs Premium comparison
        embed.add_field(
            name="🆓 Ücretsiz Özellikler",
            value="""
• Temel moderasyon (ban, kick, mute)
• 10 adet custom komut
• Basik log sistemi
• Temel istatistikler
            """.strip(),
            inline=True
        )
        
        embed.add_field(
            name="💎 Premium Özellikler",
            value="""
• 🤖 AI Assistant & Analytics
• 📊 Gelişmiş dashboard
• 🎮 Oyun sistemleri
• 🎵 Müzik botu
• 🎫 Ticket sistemi
• 📈 Detaylı raporlar
• ⚡ Priority support
• 🔧 Custom integrations
            """.strip(),
            inline=True
        )
        
        # Pricing
        embed.add_field(
            name="💰 Premium Fiyatlandırma",
            value="""
**Starter:** $9.99/ay
• AI Assistant
• Analytics Dashboard
• 50 custom komut

**Pro:** $19.99/ay
• Tüm Starter özellikleri
• Müzik botu
• Ticket sistemi
• Priority support

**Enterprise:** $49.99/ay
• Tüm özellikler
• Custom integrations
• Dedicated support
• Custom branding
            """.strip(),
            inline=False
        )
        
        embed.add_field(
            name="🚀 Premium Satın Al",
            value="[Premium Satın Al](https://your-website.com/premium)\n[Demo Talep Et](https://your-website.com/demo)",
            inline=False
        )
        
        await ctx.send(embed=embed)
    
    @commands.command()
    async def premium_ai(self, ctx, *, question=None):
        """Premium AI features (example)."""
        if not self.is_premium_server(ctx.guild.id):
            embed = create_embed(
                title="💎 Premium Özellik",
                description="Bu özellik sadece Premium aboneleri için! `!premium` komutu ile detayları öğrenin.",
                color=discord.Color.gold()
            )
            return await ctx.send(embed=embed)
        
        # Premium AI features here
        embed = create_embed(
            title="🤖 Premium AI Assistant",
            description="Bu premium özellik aktif olacak!",
            color=discord.Color.green()
        )
        await ctx.send(embed=embed)

async def setup(bot):
    await bot.add_cog(PremiumFeatures(bot))
