pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "pynacl>=1.5.0",
    "requests>=2.32.4",
]
